vlan endpoint support added
currently vlan endpoint support is not there. this pr is a step
towards vlan implementation to allow vlan based CNI plugins
with kata container runtime

Signed-off-by: Saurabh Deoras <saurabh@diamanti.com>

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch 1.8.0-patched
Changes to be committed:
	modified:   virtcontainers/endpoint.go
	modified:   virtcontainers/network.go
	new file:   virtcontainers/vlan_endpoint.go